Day_9/financial_advisor/api/requirements.txt
FastAPI>=0.109.0
uvicorn[standard]>=0.27.0
yfinance>=0.2.36
httpx>=0.26.0
python-dotenv>=1.0.0
